requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
pypdf>=3.17.0
pydantic>=2.5.0
python-dotenv>=1.0.0

